Create `@webpack-blocks/webpack-common` package to make webpack & webpack2 DRY + add webpack instance to context
