+++ packages/webpack-common/.gitignore
@@ -1,0 +1,3 @@
+lib/
+node_modules/
+npm-debug.log
