+++ packages/webpack-common/README.md
@@ -1,0 +1,15 @@
+# Webpack blocks - Webpack shared functionality
+
+[![JavaScript Style Guide](https://img.shields.io/badge/code%20style-standard-brightgreen.svg)](http://standardjs.com/)
+[![NPM Version](https://img.shields.io/npm/v/@webpack-blocks/webpack.svg)](https://www.npmjs.com/package/@webpack-blocks/webpack-common)
+
+This is the `webpack-common` package. It contains code that can be shared between different webpack versions' webpack-block packages in order to make them DRY.
+
+
+## Webpack blocks
+
+Check out the
+
+👉 [Main Documentation](https://github.com/andywer/webpack-blocks)
+
+Released under the terms of the MIT license.
