+++ packages/webpack-common/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@webpack-blocks/webpack-common",
+  "version": "0.3.0",
+  "description": "Code that can be shared between different webpack versions' webpack-block packages.",
+  "license": "MIT",
+  "author": "Andy Wermke <andy@dev.next-step-software.com>",
+  "main": "lib/index",
+  "scripts": {
+    "build": "babel src/ -d lib/ --source-maps",
+    "test": "standard && ava src/**/__tests__",
+    "prepublish": "npm run build && npm test"
+  },
+  "ava": {
+    "require": [
+      "babel-register"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "es2015-node4"
+    ]
+  },
+  "engines": {
+    "node": ">= 4.0"
+  },
+  "repository": "andywer/webpack-blocks",
+  "bugs": "https://github.com/andywer/webpack-blocks/issues",
+  "devDependencies": {
+    "ava": "^0.16.0",
+    "babel-cli": "^6.16.0",
+    "babel-preset-es2015-node4": "^2.1.1",
+    "babel-register": "^6.18.0",
+    "standard": "^8.1.0",
+    "@webpack-blocks/core": "^0.1.2"
+  }
+}
