+++ packages/webpack-common/src/index.js
@@ -1,0 +1,2 @@
+export { default as defineConstants } from './defineConstants'
+export * from './webpack'
