+++ packages/webpack2/package.json
@@ -20,6 +20,7 @@
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "webpack": "^2.1.0-beta.27",
+    "@webpack-blocks/webpack-common": "^0.3.0",
     "@webpack-blocks/core": "^0.1.2"
   },
   "devDependencies": {
